fix: silence -Wconversion in Axis/BinningData/IMultiAxis for traccc
Dropping -isystem exposes these Core headers to traccc's clang
-Wconversion/-Wsign-conversion -Werror build (traccc_build_sycl_intel).
Add explicit casts at the signed/unsigned and integer/floating bin-index
boundaries:
- Axis: bin-index arithmetic in neighborHoodIndices (Open/Bound/Closed),
  wrapBin returns, and getBin{Lower,Upper}Bound/getBinCenter width math.
- IMultiAxis: iterator advance/difference over the size_t index.
- BinningData: the searchGlobal clamp ternary.

All casts are behaviour-preserving (indices are non-negative and within
int range). No functional change.

+++ Core/include/Acts/Utilities/BinningData.hpp
@@ -591,7 +591,7 @@ class BinningData {
             : bin;
     return static_cast<std::size_t>(
         (bin <= static_cast<int>(bData.m_bins - 1))
-            ? bin
+            ? static_cast<std::size_t>(bin)
             : ((bData.option == open) ? (bData.m_bins - 1) : 0));
   }
